pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sing"
version = "0.1.0"
description = "SING v1 single-image semantic invariant analysis."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "torch>=2.1",
  "torchvision>=0.16",
  "timm>=1.0.0",
  "numpy>=1.24",
  "pillow>=10.0",
  "pyyaml>=6.0",
  "diffusers>=0.30",
  "transformers>=4.40,<5",
  "accelerate>=0.30",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
sing = "sing.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
